+++ velocidi-sdk/src/main/kotlin/com/velocidi/events/OrderPlace.kt
@@ -1,0 +1,16 @@
+package com.velocidi.events
+
+data class OrderPlace(
+    override val siteId: String,
+    override val clientId: String
+) : TrackingEvent("orderPlace") {
+    /**
+     * Line items ordered
+     */
+    var lineItems: List<LineItem> = emptyList()
+
+    /**
+     * Order details
+     */
+    var order: Order? = null
+}